@@ -3,7 +3,7 @@
     "description": "An mbstring stream conversion filter.",
     "type": "library",
     "require-dev": {
-        "phpunit/phpunit": ">=6.3 <8"
+        "phpunit/phpunit": "^9.5"
     },
     "license": "MIT",
     "authors": [
@@ -13,7 +13,8 @@
         }
     ],
     "require": {
-        "php": ">=7.0"
+        "php": "8.1.x",
+        "ext-mbstring": "*"
     },
     "autoload": {
         "classmap": ["src/"]
